+++ Source/Core/Common/HostDisassembler.cpp
@@ -1,7 +1,7 @@
 // Copyright 2008 Dolphin Emulator Project
 // SPDX-License-Identifier: GPL-2.0-or-later
 
-#include "UICommon/Disassembler.h"
+#include "Common/HostDisassembler.h"
 
 #include <sstream>
