+++ Source/Core/DolphinLib.props
@@ -117,6 +117,7 @@
     <ClInclude Include="Common\GL\GLX11Window.h" />
     <ClInclude Include="Common\Hash.h" />
     <ClInclude Include="Common\HookableEvent.h" />
+    <ClInclude Include="Common\HostDisassembler.h" />
     <ClInclude Include="Common\HRWrap.h" />
     <ClInclude Include="Common\HttpRequest.h" />
     <ClInclude Include="Common\Image.h" />
@@ -549,7 +550,6 @@
     <ClInclude Include="InputCommon\KeyboardStatus.h" />
     <ClInclude Include="UICommon\AutoUpdate.h" />
     <ClInclude Include="UICommon\CommandLineParse.h" />
-    <ClInclude Include="UICommon\Disassembler.h" />
     <ClInclude Include="UICommon\DiscordPresence.h" />
     <ClInclude Include="UICommon\GameFile.h" />
     <ClInclude Include="UICommon\GameFileCache.h" />
@@ -809,6 +809,7 @@
     <ClCompile Include="Common\GL\GLInterface\WGL.cpp" />
     <ClCompile Include="Common\GL\GLUtil.cpp" />
     <ClCompile Include="Common\Hash.cpp" />
+    <ClCompile Include="Common\HostDisassembler.cpp" />
     <ClCompile Include="Common\HRWrap.cpp" />
     <ClCompile Include="Common\HttpRequest.cpp" />
     <ClCompile Include="Common\Image.cpp" />
@@ -1207,7 +1208,6 @@
     <ClCompile Include="InputCommon\InputProfile.cpp" />
     <ClCompile Include="UICommon\AutoUpdate.cpp" />
     <ClCompile Include="UICommon\CommandLineParse.cpp" />
-    <ClCompile Include="UICommon\Disassembler.cpp" />
     <ClCompile Include="UICommon\DiscordPresence.cpp" />
     <ClCompile Include="UICommon\GameFile.cpp" />
     <ClCompile Include="UICommon\GameFileCache.cpp" />
